requirements.txt
Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

ML Libraries
sentence-transformers==2.2.2
spacy==3.7.2
faiss-cpu==1.7.4
torch==2.1.0

Data Processing
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.0
openpyxl==3.1.2

Utilities
tqdm==4.66.0
requests==2.31.0